Change bezant to look like other roundels.
The user guide indicates that a bezant is simply a roundel, like a golpe, guze,
hurt, and so on. This matches other sources, which describe a bezant as a
roundel or. So it is surprising if "bezant" is different from "roundel or". Most
illustrations including bezants also show them as flat discs, not as
three-dimensional representations of coins in perspective.

Tested:
$ php drawshield.php 'Azure, three bezants'