Case-insensitivity for symbols.
For alchemical, astronomical, and zodiac symbols, and names of fonts, a token
from the blazon is used as a file name. This doesn't work if the token is
capitalised, because the file names are lower-case. Case-insensitivity is
especially important for astronomical and zodiac symbols, because the names
there (e.g., Saturn, Capricorn) are properly written with an initial capital
letter.

Tested:
(alchemical/symbol) Argent, an alchemical symbol for Air sable
(astronomical/symbol) Argent, an astronomical symbol for Venus sable
(characters/letter) Argent, the letter K in a Script font sable
(symbol/any) Argent, a symbol for Water sable
(symbol/any) Argent, a symbol for Jupiter sable
(zodiac/symbol) Argent, a zodiac symbol for Leo sable