Fix furs on inverted objects.
The previous code attempted to access an array element by numeric index, when
it's actually an associative array. This caused a warning from PHP and a
malformed `translate` transform in the SVG.

Test blazon: Argent, a chevron inverted pean.

Before: The expression `$fur_data[1]` caused the following notice from PHP:
PHP Notice:  Undefined offset: 1 in /home/brian/repos/Drawshield-Code/svg/tinctures.inc on line 179
And the output SVG contained the following:
  <g transform="translate(0,) scale(1,-1)">
Note the malformed `translate` here. The result was that the ermine spots were
either upside-down or the fur was not drawn at all, depending on the SVG
renderer.

After: The shield is drawn correctly, with the ermine spots upright.