Fix "*-bundle-3 not drawn with 3" warnings.
Test blazons:

Argent, a bundle of tulips gules.
Argent, a bunch of grapes vert.

Previously these produced the following warnings, respectively:

WARNING tulip-bundle-3 not drawn with 3
WARNING grape-bundle-3 not drawn with 3

These warnings are confusing, because the blazons don't specify a number; and in
the case of the tulips, the bundle *is* drawn with 3 tulips.

The cause of the warnings is that the grammar assigns the default number 3 to
bundling words like "bundle" or "bunch", after which it seems as if the number
has been specified in the blazon.

The fix to silence these warnings is to rename the SVG files to include "-3".

I'm pretty confident this is the right solution for the tulips. It's less clear
that it's the right solution for the grapes, because there is no real sense in
which this SVG is a "bundle of 3 grapes". The advantage of this quick fix is
that no warning is produced for a blazon specifying "a bunch of grapes", and it
seems unlikely that a blazon would attempt to specify the precise number of
grapes in the bunch ("a bunch of 20 grapes"?).