Rename armour/helmet-closed to armour/helmet-knight-closed.
The visual catalog (https://drawshield.net/catalog/charges/armour/index.html)
says that this charge is "Helmet Closed". But the blazon "Argent, a helmet
closed gules" produces the warning "helmet-knight-closed not drawn with knight".
This is because the parser interprets "helmet" as "helmet-knight", which is then
modified by "closed". There seems to be no way to get this charge without this
warning.

This is certainly not the same style of helmet as armour/helmet-knight, so
perhaps this requires some further thought.

Test blazon: Argent, a helmet closed gules.