Fix Heneage knot.
The previous regex accepted the spellings "Hanage", "Haneage", "Heanage", and
"Heaneage", but not the correct spelling "Heneage". Google returns no results
for the quoted phrases "hanage knot" and "haneage knot", so it seems likely that
the original intent was to support the spellings "Heanage", "Heaneage",
"Henage", and "Heneage".

Test blazon: Argent, a Heneage knot sable.