Fix "{carp,catfish} haurient embowed".
These charges previously had the "haurient" built into the SVG. But haurient is
implemented in general by rotating the charge, so the end result was that these
fish were lying on their backs.

This change rotates the SVG images 90 degrees counterclockwise and removes
"haurient" from the file names.

Test blazons:
Or, a carp haurient embowed sable.
Or, a catfish haurient embowed sable.